requirements.txt
fastapi
uvicorn
python-multipart
jinja2
google-cloud-storage
pandas
google-auth
pandas-gbq
